requirements.txt
bs4
pdf2image
pillow
python-dotenv
google-genai
mistralai
